Major changes to motion models
Changed motion model interface to include setter helpers. They take as input state vectors and directly modify them. This follows the paradigm that outsiders should not need to know how the model states are arranged to use the models.

Motion models now expose the names of the input signals. Outsiders can use it to construct the input dictionary or make modifications to an existing input dictionary (like adding noise to a control signal). Again, outsiders do not need to know the arrangement of input signals in the models, only matching them in a dictionary based on the names.

Changed DifferentialDriveTorqueInput to use [v,w] as state instead of wheel angular velocity.
This is more consistent with how the (cascaded) controllers view the (cascaded) system, which make it easier to set/interpret initial and goal states.

Wrote comprehensive tests for controllers and estimators, covering the subsystems as well as the cascaded full system.

Next step is to implement the Particle Filter based pose estimator.